# Avoid excessive property verifications when adding resource properties in OData reader