--- a/src/Microsoft.OData.Core/JsonLight/ODataJsonLightPropertyAndValueDeserializer.cs
+++ b/src/Microsoft.OData.Core/JsonLight/ODataJsonLightPropertyAndValueDeserializer.cs
@@ -1024,17 +1024,15 @@ namespace Microsoft.OData.JsonLight
             ODataResourceBase resource = resourceState.Resource;
             Debug.Assert(resource != null, "resource != null");
 
-            // The resource.Properties setter would run property verification
-            // each time a property is added to the resource. That is
-            // expensive and scales quadratically with the number of properties + number of items in collection properties.
-            // To avoid this cost, we skip property verification.
+            // To avoid repeated computations for the property verification logic inside the
+            // resource.Properties setter, we update the resource.Properties in-place
+            // without invoking the setter each time.
             // Since the resource and properties are created internally by the reader,
             // we don't need to verify the properties if we can guarantee that we are only
             // adding properties with acceptable values (i.e. not ODataResourceValue).
             // We use debug-time assertions instead to catch such bugs in tests.
 
-            // This is a stop-gap measure to address a performance issue. But
-            // we should rethink the current approach to property verification.
+            // We should rethink the current approach to property verification.
             // See: https://github.com/OData/odata.net/issues/3263
 
             Debug.Assert(!(property.Value is ODataResourceValue), Strings.ODataResource_PropertyValueCannotBeODataResourceValue(property.Name));
@@ -1042,12 +1040,14 @@ namespace Microsoft.OData.JsonLight
                 !(property.Value is ODataCollectionValue collectionValue && collectionValue.Items.Any(item => item is ODataResourceValue)),
                 Strings.ODataResource_PropertyValueCannotBeODataResourceValue(property.Name));
 
-            bool originalSkipPropertyVerificationValue = resource.SkipPropertyVerification;
-            resource.SkipPropertyVerification = true;
-            
-            resource.Properties = resource.Properties.ConcatToReadOnlyEnumerable("Properties", property);
+            if (resource.Properties.IsEmptyReadOnlyEnumerable())
+            {
+                resource.Properties = new ReadOnlyEnumerable<ODataProperty>();
+            }
 
-            resource.SkipPropertyVerification = originalSkipPropertyVerificationValue;
+            // Despite the name the resource.Properties type here is not actually
+            // readonly, it supports appending in-place.
+            resource.Properties.ConcatToReadOnlyEnumerable("Properties", property);
             return property;
         }