<!-- markdownlint-disable MD002 MD041 -->

### Issues

*This pull request addresses #3264 in ODL 7.x*

If this PR is approved, I'll port the fix to ODL 8.x

### Description

To fix the performance issue caused by excessively running property verification each time the reader adds a property to the resource being deserialized, I update the `resource.Properties` in-place in `ODataJsonLightPropertyAndValueDeserializer.AddResourceProperty()` without invoking the setter each time since the verification logic runs in the setter.

In lieu of running the verification logic, I use `Debug.Assert` statements to perform the equivalent verification during debug builds only, to avoid any runtime impact from the verification. I could have implemented the verification using logic that's executed also in prod builds that would also thrown an exception, this would still improve performance since the properties would be checked in O(n) times instead of O(n^2) times, but I found this unnecessary.

Here's the reasoning for doing debug-time verification only: If the reader stores the `ODataResourceValue` instances inside a resource's properties, then we would be getting exceptions already due to the existing property verification that runs by default. The fact that we don't get exceptions thrown in the reader related to this issue, means that the reader does not add such values. In any case, it would be a bug for the reader to add values in the resource properties which are considered invalid. Which is why I believe a `Debug.Assert` would suffice so we could catch such bugs when testing. Since these resources and properties are created inside the reader, we have full control of the internal behaviour.

The reason I chose this approach (over other alternatives listed in the attached issue) is because it's simple and localized to a single internal method. Every user will benefit from this without having to change their code or opt-in through some new flag. This is a non-breaking change. And we can easily change this implementation in the future without breaking.

Again, I do suggest we revisit the property verification and decide how to address it in a more holistic way since it also affects the writer performance and may affect other scenarios as well (see: https://github.com/OData/odata.net/issues/3263)

### Benchmarks

Benchmark code: 

#### Before

```sh
// * Summary *

BenchmarkDotNet v0.15.1, Windows 11 (10.0.26100.3775/24H2/2024Update/HudsonValley)
Intel Xeon W-2123 CPU 3.60GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK 10.0.100-preview.1.25120.13
  [Host]     : .NET 8.0.17 (8.0.1725.26602), X64 RyuJIT AVX-512F+CD+BW+DQ+VL
  DefaultJob : .NET 8.0.17 (8.0.1725.26602), X64 RyuJIT AVX-512F+CD+BW+DQ+VL
```


| Method       | Mean     | Error    | StdDev   | Gen0      | Gen1     | Gen2    | Allocated |
|------------- |---------:|---------:|---------:|----------:|---------:|--------:|----------:|
| ReadResource | 23.16 ms | 0.459 ms | 0.969 ms | 2562.5000 | 562.5000 | 62.5000 |  12.65 MB |

![image](https://github.com/user-attachments/assets/2419b554-c64d-4c5c-b286-772f864c5743)

#### After

```sh
// * Summary *

BenchmarkDotNet v0.15.1, Windows 11 (10.0.26100.3775/24H2/2024Update/HudsonValley)
Intel Xeon W-2123 CPU 3.60GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK 10.0.100-preview.1.25120.13
  [Host]     : .NET 8.0.17 (8.0.1725.26602), X64 RyuJIT AVX-512F+CD+BW+DQ+VL
  DefaultJob : .NET 8.0.17 (8.0.1725.26602), X64 RyuJIT AVX-512F+CD+BW+DQ+VL
```


| Method       | Mean     | Error     | StdDev    | Gen0     | Gen1     | Allocated |
|------------- |---------:|----------:|----------:|---------:|---------:|----------:|
| ReadResource | 4.377 ms | 0.0867 ms | 0.1243 ms | 609.3750 | 453.1250 |   3.26 MB |

![image](https://github.com/user-attachments/assets/739481df-7168-4c29-abf3-54b9bb650157)


### Checklist (Uncheck if it is not completed)

- [ ] *Test cases added*
- [x] *Build and test with one-click build and test script passed*

### Additional work necessary

*If documentation update is needed, please add "Docs Needed" label to the issue and provide details about the required document change in the issue.*

### Repository notes

Team members can start a CI build by adding a comment with the text `/AzurePipelines run` to a PR. A bot may respond indicating that there is no pipeline associated with the pull request. This can be ignored if the build is triggered.

Team members should **not** trigger a build this way for pull requests coming from forked repositories. They should instead trigger the build manually by setting the "branch" to `refs/pull/{prId}/merge` where `{prId}` is the ID of the PR.